[pulseaudio]: Created new plugin for pulseaudio
This commit adds a new plugin for pulseaudio, which is the default sound server
middleware package in Ubuntu LTSes up to and including Jammy as well as RedHat
up to and including RHEL 9. The plugin collects the output of commands that list
the detected sinks, sources, and soundcards, which helps determine the available
input and output audio devices. The other commands included aid in determining
any applied configurations which may affect the prioritization of device
selection.

It seemed appropriate to move the add_copy_spec of the pulseaudio directory from
the soundcard plugin to the new pulseaudio plugin for organizational structure,
grouping it with the rest of the package's collection. Especially as Ubuntu
LTSes will be using Pipewire by default from Noble onwards, so PulseAudio is no
longer representative of the default audio/soundcard configuration

Signed-off-by: Bryan Fraschetti bryan.fraschetti@canonical.com