[microk8s: Include files for debugging dqlite
This commit adds some configuration files that store information about the
underlying dqlite cluster that backs microk8s. At the same time it runs several
queries against the db to determine the status of other nodes (essentially
collecting the worldview of this node). It collects a directory listing of the
database directory to ensure consistency amongst closed intervals.