[dqlite] Created new plugin for dqlite
This commit adds a new plugin for dqlite which can be triggered by the
several projects that leverage its capabilities such as microcloud,
microceph, microovn, microk8s, and lxd. As it is an embedded db, each
application uses it slightly differently - ie. different tables,
different columns, etc. Common collection points between the activators
are placed in the baseCollection function while application-specific
collections go under the eponymous collector functions, some of which
are no-ops as of right now. The command output is stored under
dqlite/{activator}/ to categorize and organize the collection since one
or more of the activators may be present on the system and each should
be collected.

The setup function runs the general collection and the
application-specific collector to generate a list of applicable queries
to run against dqlite. For most of the activators, these are then run
against the db in two ways - using the built in commands some
applications offer (eg. microceph cluster sql) as well as by curling the
unix socket. These are then executed in the run_batch_dqlite_queries
function.

Signed-off-by: Bryan Fraschetti <bryan.fraschetti@canonical.com>